requirements.txt
alembic==1.12.1
astroid==3.0.1
blinker==1.7.0
click==8.1.7
coverage==7.3.2
dill==0.3.7
Flask==3.0.0
Flask-Alembic==2.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
platformdirs==4.0.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pylint==3.0.2
pytest==7.4.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
tomlkit==0.12.3
typing_extensions==4.8.0
Werkzeug==3.0.1
